pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "crawlhog"
dynamic = ["version"]
description = "A robust documentation web scraper"
readme = "README.md"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Documentation",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "click>=8.0.0",
    "firecrawl-py>=1.11.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/crawlhog"
Repository = "https://github.com/yourusername/crawlhog.git"

[project.scripts]
chog = "crawlhog.cli.main:cli" 